Update:ring_buffer
Added changes and fixed some problems
